+++ internal/repositories/capa5repo/gormrepo/unit.go
@@ -1,0 +1,4 @@
+package gormrepo
+
+type unit struct {
+}