+++ web/capa5/types/unit.ts
@@ -1,13 +1,37 @@
+export enum UnitKind {
+    Group = "group",
+    Squadron = "squadron",
+    Flight = "flight",
+}
+
+export enum UnitCategory {
+    Admin = "admin",
+    Composite = "composite",
+    Cadet = "cadet",
+    Senior  = "senior",
+}
+
 export class Unit {
-    name: string;
     charterNumber: string;
+    kind: UnitKind;
+    category: UnitCategory;
+    name: string;
     address: string;
     city: string;
 
-    constructor(name: string, charterNumber: string, address: string, city: string) {
-        this.name = name;
+    constructor(
+        charterNumber: string,
+        kind: string,
+        category: string,
+        name: string,
+        address: string,
+        city: string,
+    ) {
         this.charterNumber = charterNumber;
-        this.address = address;
-        this.city = city;
+        this.kind = kind
+        this.category = category
+        this.name = name
+        this.address = address
+        this.city = city
     }
 }