+++ pkg/units/CHANGELOG.md
@@ -8,8 +8,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added `Wing.RegionLookup()`
+- Added `ParseShortCharter()`
 
 ### Changed
+- `UnitCharterNumber` now implements the `json.Marshaller` and `json.Unmarshaller` interfaces
+- `UnitCharterNumber` now implements the `sql.Scanner` and `sql.Valuer` interfaces
+- `UnitKind` now implements the `json.Marshaller` and `json.Unmarshaller` interfaces
+- `UnitKind` now implements the `sql.Scanner` and `sql.Valuer` interfaces
+- `UnitCategory` now implements the `json.Marshaller` and `json.Unmarshaller` interfaces
+- `UnitCategory` now implements the `sql.Scanner` and `sql.Valuer` interfaces
 
 ### Deprecated
 
@@ -22,10 +30,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v1.0.0]
 
 ### Added
--   Copied over the unit types/functionality from v2.2.0 of the [old library](https://github.com/ag7if/cap).
+- Copied over the unit types/functionality from v2.2.0 of the [old library](https://github.com/ag7if/cap).
 
 ### Removed
--   `MemberType` and `Grade` from v2.2.0 of the [old library](https://github.com/ag7if/cap) moved to the `members`
+- `MemberType` and `Grade` from v2.2.0 of the [old library](https://github.com/ag7if/cap) moved to the `members`
     library.
 
 [unreleased]: https://github.com/DerHabicht/cap-tools/compare/units-v1.0.0...HEAD